requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
mlflow
dvc
fastapi
uvicorn
pydantic
python-dotenv
joblib
pyarrow
pytest
requests

